bot/pyproject.toml
[tool.poetry]
name = "discordbot"
version = "0.1.0"
description = ""
authors = ["Baranov_DR <mrramzes37@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
discord = "^2.3.1"
cachetools = "^5.3.1"
python-dotenv = "^1.0.0"
fastapi = "^0.100.0"
lavalink = "^4.0.6"
httpx = "^0.24.1"
pytest = "^7.4.0"
pydantic-sqlalchemy = "^0.0.9"
sqlmodel = "^0.0.8"
pillow = "^10.0.0"
jinja2 = "^3.1.2"
table2ascii = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
